+++ backend/domain/product.go
@@ -6,6 +6,23 @@ import (
 	"github.com/PUArallelepiped/PUN-street-Universal-Access/swagger"
 )
 
+type PostInfo struct {
+	Key    string `json:"key"`
+	Action string `json:"action"`
+	Source string `json:"source"`
+	Format string `json:"format"`
+}
+
+type ResponseInfo struct {
+	StatusCode int       `json:"status_code"`
+	Image      ImageInfo `json:"image"`
+}
+
+type ImageInfo struct {
+	Url       string `json:"url"`
+	UrlViewer string `json:"url_viewer"`
+}
+
 type ProductRepo interface {
 	GetByID(ctx context.Context, id int64) (*[]swagger.ProductInfo, error)
 	GetByProductID(ctx context.Context, id int64) (*swagger.ProductInfo, error)
